+++ RayTracerChallenge/Source/Vectors/Point.cs
@@ -2,7 +2,7 @@
 {
     public class Point : Tuple
     {
-        public Point(float x, float y, float z) : base(x, y, z)
+        public Point(double x, double y, double z) : base(x, y, z)
         {
             this.x = x;
             this.y = y;
