+++ RayTracerChallenge/Source/Vectors/Tuple.cs
@@ -4,10 +4,10 @@ namespace RT.Source.Vectors
 {
     public class Tuple
     {
-        public float x;
-        public float y;
-        public float z;
-        public float w;
+        public double x;
+        public double y;
+        public double z;
+        public double w;
 
         public Tuple()
         {
@@ -18,14 +18,14 @@ namespace RT.Source.Vectors
         }
 
         // For Point/Vector inheritance
-        public Tuple(float x, float y, float z)
+        public Tuple(double x, double y, double z)
         {
             this.x = x;
             this.y = y;
             this.z = z;
         }
 
-        public Tuple(float x, float y, float z, float w)
+        public Tuple(double x, double y, double z, double w)
         {
             this.x = x;
             this.y = y;
@@ -45,9 +45,9 @@ namespace RT.Source.Vectors
 
         public static Tuple operator -(Tuple a, Tuple b) => new(a.x - b.x, a.y - b.y, a.z - b.z, a.w - b.w);
 
-        public static Tuple operator *(Tuple a, float b) => new(a.x * b, a.y * b, a.z * b, a.w * b);
+        public static Tuple operator *(Tuple a, double b) => new(a.x * b, a.y * b, a.z * b, a.w * b);
 
-        public static Tuple operator /(Tuple a, float b) => new(a.x / b, a.y / b, a.z / b, a.w / b);
+        public static Tuple operator /(Tuple a, double b) => new(a.x / b, a.y / b, a.z / b, a.w / b);
 
         public static bool operator ==(Tuple a, Tuple b)
         {
@@ -69,17 +69,17 @@ namespace RT.Source.Vectors
 
         #region Transformations
 
-        public Tuple Translate(float x, float y, float z, bool inverse = false) => Matrix.Translation(x, y, z, inverse) * this;
+        public Tuple Translate(double x, double y, double z, bool inverse = false) => Matrix.Translation(x, y, z, inverse) * this;
 
-        public Tuple Scale(float x, float y, float z, bool inverse = false) => Matrix.Scaling(x, y, z, inverse) * this;
+        public Tuple Scale(double x, double y, double z, bool inverse = false) => Matrix.Scaling(x, y, z, inverse) * this;
 
-        public Tuple RotateX(float rad, bool inverse = false) => Matrix.RotationX(rad, inverse) * this;
+        public Tuple RotateX(double rad, bool inverse = false) => Matrix.RotationX(rad, inverse) * this;
 
-        public Tuple RotateY(float rad, bool inverse = false) => Matrix.RotationY(rad, inverse) * this;
+        public Tuple RotateY(double rad, bool inverse = false) => Matrix.RotationY(rad, inverse) * this;
 
-        public Tuple RotateZ(float rad, bool inverse = false) => Matrix.RotationZ(rad, inverse) * this;
+        public Tuple RotateZ(double rad, bool inverse = false) => Matrix.RotationZ(rad, inverse) * this;
 
-        public Tuple Skew(float xy, float xz, float yx, float yz, float zx, float zy, bool inverse = false)
+        public Tuple Skew(double xy, double xz, double yx, double yz, double zx, double zy, bool inverse = false)
             => Matrix.Skewing(xy, xz, yx, yz, zx, zy, inverse) * this;
 
         #endregion
